+++ BTDB/IOC/IContanerRegistration.cs
@@ -3,4 +3,5 @@ namespace BTDB.IOC;
 interface IContanerRegistration
 {
     void Register(ContainerRegistrationContext context);
+    void RegisterForServiceCollection(ServiceCollectionRegistrationContext context);
 }