+++ BTDB/IOC/ServiceCollectionExtensions.cs
@@ -1,0 +1,29 @@
+using System.Collections.Generic;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace BTDB.IOC;
+
+public static class ServiceCollectionExtensions
+{
+    public static IServiceCollection UseBtdbIoc(this IServiceCollection services, ContainerBuilder containerBuilder,
+        ContainerVerification options = ContainerVerification.AllTypesAreGenerated)
+    {
+        var builderServices = containerBuilder.GetServiceCollection();
+        if (builderServices != null)
+        {
+            foreach (var descriptor in builderServices)
+            {
+                ((ICollection<ServiceDescriptor>)services).Add(descriptor);
+            }
+        }
+
+        var integration = new ServiceProviderIntegration();
+        var exports = containerBuilder.CollectServiceCollectionRegistrations().Registrations;
+        integration.RegisterExternalServices(services, exports, (serviceProvider, serviceProviderIntegration) =>
+        {
+            containerBuilder.SetServiceProvider(serviceProvider, serviceProviderIntegration);
+            return (ContainerImpl)containerBuilder.BuildAndVerify(options);
+        });
+        return services;
+    }
+}